Deserialization test + adding GPU to report

+++ RackPeek.Domain/Resources/Hardware/Models/AccessPoint.cs
@@ -3,5 +3,5 @@ namespace RackPeek.Domain.Resources.Hardware.Models;
 public class AccessPoint : Hardware
 {
     public string? Model { get; set; }
-    public int? Speed { get; set; }
+    public double? Speed { get; set; }
 }
